Extract RowAtlas into its own file
